requirements.txt
black
Django
django-oauth-toolkit
django-cors-headers
django-filter
django-guardian
djangorestframework
gunicorn
requests
